moved computations of effective strain rate to different files to solve compilation issues.
